requirements.txt
# 图像处理和计算机视觉
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# 屏幕截图和窗口操作
pyautogui>=0.9.54
mss>=9.0.1
# pygetwindow在macOS上有兼容性问题，我们将使用替代方案

# OCR文字识别
easyocr>=1.7.0

# 网络请求
requests>=2.31.0
aiohttp>=3.9.0

# 日志和配置
pyyaml>=6.0.0
python-dotenv>=1.0.0

# 开发和调试工具
pytest>=7.4.0
black>=23.11.0 